requirements.txt
asgiref==3.6.0
certifi==2022.12.7
cffi==1.17.1
charset-normalizer==3.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==45.0.3
dj-database-url==2.1.0
Django==4.2.7
django-axes==6.3.0
django-cors-headers==4.3.0
django-filter==23.3
django-password-validators==1.3.0
django-ratelimit==4.1.0
django-rest-knox==4.2.0
django-storages==1.14.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
dotenv==0.9.9
gunicorn==21.2.0
idna==3.4
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.2
packaging==25.0
Pillow==10.1.0
psycopg2-binary==2.9.9
pycparser==2.22
PyJWT==2.10.1
python-dotenv==1.0.0
pytz==2022.7.1
requests==2.28.2
sqlparse==0.4.3
typing_extensions==4.14.0
tzdata==2025.2
uritemplate==4.1.1
urllib3==1.26.15
whitenoise==6.5.0
yapf==0.32.0
